KE-5550, add config 'kylin.job.delete-job-tmp-when-retry' to control whether delete job_tmp when job retry or not
* KE-5550, add config 'kylin.job.delete-job-tmp-when-retry' to control whether delete job_tmp when job retry or not

* fix ut

* retain driver log by-default

* minor, remove unused code

+++ src/spark-project/engine-spark/src/main/java/io/kyligence/kap/engine/spark/job/NSparkExecutable.java
@@ -192,7 +192,7 @@ public class NSparkExecutable extends AbstractExecutable {
             jars = kylinJobJar;
         }
 
-        if (!isResumable()) {
+        if (!isResumable() && config.isDeleteJobTmpWhenRetry()) {
             deleteJobTmpDirectoryOnExists();
         }
